MGDSTRM-8524: move rule related class into rule package
