+++ src/main/java/io/bf2/kafka/common/Config.java
@@ -3,5 +3,5 @@ package io.bf2.kafka.common;
 public class Config {
     public static final String PREFIX = "kas.";
 
-    static final String POLICY_PREFIX = PREFIX + "policy.";
+    public static final String POLICY_PREFIX = PREFIX + "policy.";
 }
